==> replication_data/rmd/02_format_and_clean.rmd <==
---
title: "02_format_and_clean"
output:
  html_document:
    df_print: paged
  toc: yes
editor_options:
  chunk_output_type: inline
---


```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)

`%>%` <- magrittr::`%>%`

```

This assigns sentence numbers by the text spans instead of the original gui

```{r}

icb_long_spans <- readRDS(file=paste0(here::here(),"/replication_data/in/icb_long_spans.Rds"))

coders_per_crisis <- icb_long_spans %>% dplyr::select(crisno,  email_id, expert) %>% distinct() #%>% dplyr::count(crisno)
coders_per_crisis_tally <- coders_per_crisis %>% mutate(novice=expert==0) %>% group_by(crisno) %>% summarise( expert=sum(expert==1), novice=sum(novice==1) )
crisis_sentences_per_coder <- icb_long_spans %>% filter(varname %>% str_detect("sentence_events")) %>% dplyr::count(email_id, crisno)
    
    
dim(icb_long_spans) #767100     16
icb_long_spans_clean <- icb_long_spans %>%
      filter(!is.na(sentence_span_start) & sentence_span_score>0.35 & sentence_clean!='PRE-CRISIS') %>%
      group_by(crisno) %>%
        mutate(sentence_number_int_aligned=sentence_span_start %>% as.factor() %>% as.integer() ) %>%
      ungroup() %>%
      filter(!varname %in% c('sentence_number_int', 'raw_crisis', 'raw_crisno', 'raw_sentence', 'savefile_crc32', 'raw_crisno_sentno', 'raw_crisno_version_sentno', 'crisisno_crisis_version')) %>%
      mutate(varname=varname %>% recode(act_descalate = 'act_deescalate', interact_descalate = 'interact_deescalate', condition_act_descalate = 'condition_act_deescalate', 
                                        condition_interact_descalate = 'condition_interact_deescalate' ) ) %>%
  filter(!varname %in% c("sentence","text")) %>% 
  dplyr::mutate(value = strsplit(as.character(value), ";")) %>% 
  tidyr::unnest(value) %>%
  dplyr::mutate(value=value %>% tolower %>% trimws()) %>% 
  filter(!duplicated(paste(email_id, crisno,  sentence_number_int_aligned, event_number_int, varname,value)))  %>%
  mutate(sentencenumber_original= value %>% as.numeric() ) %>% #do this out of the loop or the warnings slow everything down
  group_by(email_id, crisno, input_crisis,sentence_number_int_aligned ) %>% 
    mutate(sentencenumber_original= ifelse(varname %in%  'sentencenumber', value, NA) ) %>%
    mutate(sentencenumber_original= sentencenumber_original %>% max(na.rm=T)) %>%
  ungroup()
  
dim(icb_long_spans_clean) #715451

varname_count <- icb_long_spans_clean %>% count(varname)
value_count <- icb_long_spans_clean %>% count(value)

#dplyr::mutate(sayintkind_react = ifelse(sayintkind_react=="This Sentence", raw_sentence_number_int,
#                                                   stringr::str_replace(sayintkind_react, "Sentence ","")))
                      

#we basically don't believe 'sentence  1' that just means earlier

#Map the original sentence reorder to the new span numbers
icb_long_spans_clean %>% filter(varname %in% c('do_timing','do_timing','think_sentence_events','sayintkind_react','do_timing_reorder')) %>% count(varname, value) %>% View()

alter_allignment <- icb_long_spans_clean %>% 
                    dplyr::select(crisno, email_id, input_crisis, sentence_number_int_aligned,sentencenumber_original) %>% distinct() 


reorder_normalized <- icb_long_spans_clean %>% 
                      filter(varname %in% c('do_timing','do_timing','think_sentence_events','sayintkind_react','do_timing_reorder')) %>%
                      dplyr::select(crisno, email_id, input_crisis, varname, value,sentence_number_int_aligned,sentencenumber_original) %>%
                      mutate(value_normalized = value %>% str_replace("sentence  |sentence ","") %>% trimws() %>% as.numeric() ) %>%
                      left_join(
                        alter_allignment %>% dplyr::select(crisno , email_id ,input_crisis  ,value_normalized_new=sentence_number_int_aligned ,value_normalized=sentencenumber_original) %>%
                          mutate(value_normalized=value_normalized %>% as.numeric(), value_normalized_new= value_normalized_new %>% as.numeric() )
                        ) %>%
                      mutate(value_normalized = ifelse(!is.na(value_normalized_new) , value_normalized_new, value_normalized) ) %>%
                      dplyr::select(-value_normalized_new) %>%
                      mutate(value_normalized = ifelse(value %in% c("act in this sentence","speech in this sentence","this sentence") , sentence_number_int_aligned, value_normalized) ) %>%
                      mutate(value_normalized=value_normalized %>% as.character()) %>%
                      #reorders were broken and set to default sentence 1 so drop them without loss of info
                      mutate(value_normalized = ifelse(value %in% c("sentence  1|1") & varname %in% "do_timing_reorder" , "drop", value_normalized) ) %>%
                      filter(!is.na(value_normalized)) 
dim(reorder_normalized)

```


```{r, eval=T}
# Fix typos:

print(paste("change 'premption' to 'preemption'", icb_long_spans_clean %>% dplyr::filter(stringr::str_detect(value, "premption")==TRUE) %>% nrow(), "times."))
icb_long_spans_clean <- icb_long_spans_clean %>% dplyr::mutate(value = stringr::str_replace(value, "premption", "preemption"))

print(paste("change 'disaprove' to 'disapprove'", icb_long_spans_clean %>% dplyr::filter(stringr::str_detect(value, "disaprove")==TRUE) %>% nrow(), "times."))
icb_long_spans_clean <- icb_long_spans_clean %>% dplyr::mutate(value = stringr::str_replace(value, "disaprove", "disapprove"))

print(paste("change 'uncoperative' to 'uncooperative'", icb_long_spans_clean %>% dplyr::filter(stringr::str_detect(value, "uncoperative")==TRUE) %>% nrow(), "times."))
icb_long_spans_clean <- icb_long_spans_clean %>% dplyr::mutate(value = stringr::str_replace(value, "uncoperative", "uncooperative"))

print(paste("change 'capitol' to 'capital'", icb_long_spans_clean %>% dplyr::filter(stringr::str_detect(value, "capitol")==TRUE) %>% nrow(), "times."))
icb_long_spans_clean <- icb_long_spans_clean %>% dplyr::mutate(value = stringr::str_replace(value, "capitol", "capital"))

print(paste("change 'survailance' to 'surveillance'", icb_long_spans_clean %>% dplyr::filter(stringr::str_detect(value, "survailance")==TRUE) %>% nrow(), "times."))
icb_long_spans_clean <- icb_long_spans_clean %>% dplyr::mutate(value = stringr::str_replace(value, "survailance", "surveillance"))

print(paste("change 'disaprove' to 'disapprove'", icb_long_spans_clean %>% dplyr::filter(stringr::str_detect(value, "disaprove")==TRUE) %>% nrow(), "times."))
icb_long_spans_clean <- icb_long_spans_clean %>% dplyr::mutate(value = stringr::str_replace(value, "disaprove", "disapprove"))

#Clean up some dates
icb_long_spans_clean <- icb_long_spans_clean %>% dplyr::mutate(value = stringr::str_replace(value, "17 july 2004", "17 July 2014"))


```


# Load and apply normalizations 

Load normalization dictionaries

```{r message=FALSE}

target_file <- paste0(here::here(),"/replication_data/in/icb_manual_recoding_master_sheet.xlsx")
# in not working, run googledrive::drive_auth() and select new token

#Paper 2 will take dates and time seriously we ignore now
#What we're going to do is generate several new long variables and let it allow for agreements across coders
dictionary_dates <- readxl::read_excel(target_file, sheet="dates", col_types="text") %>%
                    dplyr::select(-coder, -notes)   %>% distinct() %>% 
                    tidyr::pivot_longer(cols=-c(date_orig)) %>% 
                    filter(!is.na(value)) 
dictionary_dates$decade <- ( dictionary_dates$value %>% stringi::stri_match_first(regex="([0-9]{3}[A-Z])") )[,2] 
dictionary_dates$year <- ( dictionary_dates$value %>% stringi::stri_match_first(regex="([0-9]{4})") )[,2] %>% as.numeric()
dictionary_dates$month <- ( dictionary_dates$value %>% stringi::stri_match_first(regex="[Y0-9]{4}-([0-9]{1,2})") )[,2] %>% as.numeric()
dictionary_dates$day <- ( dictionary_dates$value %>% stringi::stri_match_first(regex="[Y0-9]{4}-[M0-9]{1,2}-([0-9]{1,2})") )[,2] %>% as.numeric()

dictionary_dates <- dictionary_dates %>% 
                    mutate(month=ifelse(str_detect(value, "Jan|jan"),1, month)) %>%
                    mutate(month=ifelse(str_detect(value, "Feb|feb"), 2, month)) %>%
                    mutate(month=ifelse(str_detect(value, "March|mar"),3, month)) %>%
                    mutate(month=ifelse(str_detect(value, "April|apr"),4, month) )%>%
                    mutate(month=ifelse(str_detect(value, "May|may"),5, month) )%>%
                    mutate(month=ifelse(str_detect(value, "June|jun"),6, month)) %>%
                    mutate(month=ifelse(str_detect(value, "July|jul"),7, month) )%>%
                    mutate(month=ifelse(str_detect(value, "Aug|aug"),8, month) )%>%
                    mutate(month=ifelse(str_detect(value, "Sep|sep"),9, month) )%>%
                    mutate(month=ifelse(str_detect(value, "Oct|oct"),10, month) )%>%
                    mutate(month=ifelse(str_detect(value, "Nov|nov"),11, month) )%>%
                    mutate(month=ifelse(str_detect(value, "Dec|dec"),12, month) )%>%

                    #filter(value=="circa autumn 1947") %>%
                    mutate(decade = ifelse(!is.na(year), floor(as.numeric(year)/10)*10  ,  as.numeric(str_extract(decade, "[0-9]{3}"))*10  ) %>% as.numeric()  ) %>% 
                    mutate(year_part = NA ) %>%
  
                    mutate(year_part = ifelse( (!is.na(year) & is.na(month) & str_detect(value, "beginning|early|first half|second quarter|first quarter") ) | 
                                                 ( !is.na(month) & month<=4 ) , 1 , year_part ) ) %>%
  
                    mutate(year_part = ifelse( (!is.na(year) & is.na(month) & str_detect(value, "middle|summer|autumn") ) |  ( !is.na(month) & month>=5 & month<=8 ), 2 , year_part ) )  %>%
  
                    mutate(year_part = ifelse(  (!is.na(year) & is.na(month) & str_detect(value, "end") ) |  ( !is.na(month) & month>=9 )  , 3 , year_part )   ) %>% 
  
                    mutate(month_part = NA ) %>%
                    mutate(month_part = ifelse( ( !is.na(year) & !is.na(month) & is.na(day) & str_detect(value, "beginning|early|first half")) |  ( !is.na(day) & day<=9 ) , 1 , month_part ) ) %>%
                    mutate(month_part = ifelse( (!is.na(year) & !is.na(month) & is.na(day) & str_detect(value, "middle")) |  (!is.na(day) &  day>=10 & day<=20 ), 2 , month_part ) )  %>%
                    mutate(month_part = ifelse( (!is.na(year) & !is.na(month) & is.na(day) & str_detect(value, "end")) |  (!is.na(day) &  day>=21 ), 3 , month_part ) ) %>%
                    dplyr::select(date_orig,  name, ,value , decade , year ,year_part , month , month_part, day ) %>%
                    arrange(decade,year, year_part, month, month_part, day) %>%
                    mutate(varname="date") %>%
                    mutate(varname_normalized="date" %>% paste0("_",name)) %>%
                    mutate(varname_normalized = ifelse(varname_normalized %in% c('date_earliest_date',
                                                                                 'date_earliest_date2',
                                                                                 'date_point_in_time',
                                                                                 'date_point_in_time2',
                                                                                 'date_point_in_time3',
                                                                                 'date_point_in_time4',
                                                                                 'date_point_in_time5',
                                                                                 'date_start_time',
                                                                                 'date_start_time2'
                                                                                 ) , 
                                                       "date_earliest" , varname_normalized   )) %>%
                    mutate(varname_normalized = ifelse(varname_normalized %in% c('date_earliest_end_date',
                                                                                 'date_end_time',
                                                                                 'date_end_time2',
                                                                                 'date_latest_date',
                                                                                 'date_latest_date2',
                                                                                 'date_latest_start_date'),  
                                                       "date_latest" , varname_normalized   ))

dictionary_dates %>% count(varname_normalized)

dates_normalized <- dictionary_dates %>% 
                    dplyr::select(date_orig, varname_normalized, decade , year, year_part, month, month_part,   day)  %>% 
                    pivot_longer(cols=-c(date_orig, varname_normalized)) %>%
                    mutate(varname_normalized=varname_normalized %>% paste0("_",name)) %>%
                    dplyr::select(value_normalized = value , value=date_orig, varname_normalized )


dictionary_actions  <- readxl::read_excel(target_file, sheet="actions") %>% dplyr::distinct() 
dictionary_condition_actions  <- readxl::read_excel(target_file, sheet="condition_actions") %>% dplyr::distinct() 

dictionary_actors    <- readxl::read_excel(target_file, sheet="actors") %>% dplyr::select(crisno,value, value_qcode=value_normalized_wikidata_id, value_normalized=value_normalized_label) %>% dplyr::distinct() %>%
                        full_join(varname_count %>% dplyr::select(varname) %>% filter(varname %>% str_detect("actor")),  by = character())

#Locations and dates are both really hard to do
dictionary_locations <- readxl::read_excel(target_file, sheet="locations", na = c("", "NA")) %>% 
                        dplyr::select(value=interact_location_other, value_normalized=norm_name, value_qcode= norm_qcode) %>% distinct() %>%
                        full_join(varname_count %>% dplyr::select(varname) %>% filter(varname %>% str_detect("location_other")),  by = character())
dictionary_condition_locations <- readxl::read_excel(target_file, sheet="condition_locations", na = c("", "NA")) %>% 
                                  dplyr::select(value=condition_interact_location_other, value_normalized=norm_name, value_qcode= norm_qcode) %>% distinct() %>%
                        full_join(varname_count %>% dplyr::select(varname) %>% filter(varname %>% str_detect("location_other")),  by = character())



dictionary_domains <- data.frame(
  stringsAsFactors = FALSE,
           varname = c("interact_domains",
                       "interact_domains","interact_domains","interact_domains",
                       "interact_domains","condition_interact_domains",
                       "condition_interact_domains","condition_interact_domains",
                       "condition_interact_domains","condition_interact_domains"),
             value = c("land","sea","air","wmd",
                       "space","land","sea","air","wmd","space"),
  value_normalized = c("land","sea","air","wmd",
                       "space","land","sea","air","wmd","space"),
       value_qcode = c("Q183366","Q15324","Q272730",
                       "Q49394","Q4169","Q183366","Q15324","Q272730",
                       "Q49394","Q4169")
)


dictionary_units <- data.frame(
  stringsAsFactors = FALSE,
                               varname = c("interact_units",
                                           "interact_units","interact_units",
                                           "interact_units","interact_units",
                                           "interact_units","interact_units",
                                           "interact_units","interact_units",
                                           "interact_units","interact_units",
                                           "interact_units","interact_units","interact_units",
                                           "interact_units",
                                           "condition_interact_units","condition_interact_units",
                                           "condition_interact_units",
                                           "condition_interact_units",
                                           "condition_interact_units",
                                           "condition_interact_units","condition_interact_units",
                                           "condition_interact_units",
                                           "condition_interact_units",
                                           "condition_interact_units","condition_interact_units",
                                           "condition_interact_units",
                                           "condition_interact_units",
                                           "condition_interact_units","condition_interact_units"),
                                 value = c("troops","armor",
                                           "artillery","fighters","bombers",
                                           "surface ships","missiles","nuclear",
                                           "aircraft carriers","surveillance",
                                           "submarines","survailance","chemical",
                                           "satellites","biological","troops",
                                           "armor","artillery","fighters",
                                           "bombers","surface ships","missiles",
                                           "nuclear","aircraft carriers",
                                           "surveillance","submarines","survailance",
                                           "chemical","satellites",
                                           "biological"),
                      value_normalized = c("troops","armor",
                                           "artillery","fighters","bombers",
                                           "surface ships","missiles","nuclear",
                                           "aircraft carriers","surveillance",
                                           "submarines","surveillance","chemical",
                                           "satellites","biological","troops",
                                           "armor","artillery","fighters",
                                           "bombers","surface ships","missiles",
                                           "nuclear","aircraft carriers",
                                           "surveillance","submarines",
                                           "surveillance","chemical","satellites",
                                           "biological"),
                           value_qcode = c("Q4991371","Q920182",
                                           "Q64418","Q127771","Q170877","Q11446",
                                           "Q974850","Q12802","Q17205",
                                           "Q3041792","Q2811","Q3041792","Q3639228",
                                           "Q26540","Q170907","Q4991371",
                                           "Q920182","Q64418","Q127771",
                                           "Q170877","Q11446","Q974850","Q12802",
                                           "Q17205","Q3041792","Q2811",
                                           "Q3041792","Q3639228","Q26540","Q170907")
                    )

dictionary_counts <- data.frame(
  stringsAsFactors = FALSE,
           varname = c("interact_forces",
                       "interact_forces","interact_forces","interact_forces",
                       "interact_forces","interact_forces","interact_forces",
                       "condition_interact_forces","condition_interact_forces",
                       "condition_interact_forces","condition_interact_forces",
                       "condition_interact_forces","condition_interact_forces",
                       "condition_interact_forces","interact_fatalities",
                       "interact_fatalities","interact_fatalities","interact_fatalities",
                       "interact_fatalities","interact_fatalities",
                       "interact_fatalities","condition_interact_fatalities",
                       "condition_interact_fatalities","condition_interact_fatalities",
                       "condition_interact_fatalities",
                       "condition_interact_fatalities","condition_interact_fatalities",
                       "condition_interact_fatalities"),
             value = c("none","individuals","tens",
                       "hundreds","thousands","tens of thousands",
                       "hundreds of thousands","none","individuals","tens","hundreds",
                       "thousands","tens of thousands","hundreds of thousands",
                       "none","individuals","tens","hundreds","thousands",
                       "tens of thousands","hundreds of thousands","none",
                       "individuals","tens","hundreds","thousands",
                       "tens of thousands","hundreds of thousands"),
  value_normalized = c("none","individuals","tens",
                       "hundreds","thousands","tens of thousands",
                       "hundreds of thousands","none","individuals","tens","hundreds",
                       "thousands","tens of thousands","hundreds of thousands",
                       "none","individuals","tens","hundreds","thousands",
                       "tens of thousands","hundreds of thousands","none",
                       "individuals","tens","hundreds","thousands",
                       "tens of thousands","hundreds of thousands"),
       value_qcode = c("Q176799","Q176799","Q176799",
                       "Q176799","Q176799","Q176799","Q176799","Q176799",
                       "Q176799","Q176799","Q176799","Q176799","Q176799",
                       "Q176799",NA,"Q1056901","Q1056901","Q1056901",
                       "Q1056901","Q1056901","Q1056901",NA,"Q1056901","Q1056901",
                       "Q1056901","Q1056901","Q1056901","Q1056901")
)

normalization_dictionaries <- bind_rows(dictionary_actions %>% rename(varname=variable, varname_normalized=variable_normalized),
                                        dictionary_condition_actions %>% rename(varname=variable, varname_normalized=variable_normalized),
                                        dictionary_locations,
                                        dictionary_condition_locations,
                                        dictionary_units,
                                        dictionary_domains,
                                        dictionary_counts
                                        )

icb_long_spans_clean_normalized <- icb_long_spans_clean %>%
                                   #Add in dates
                                   mutate(varname_normalized=NA, value_normalized=NA, value_qcode=NA) %>%
                                   left_join(dates_normalized %>% dplyr::select(value,varname_normalized2=varname_normalized, value_normalized2=value_normalized ) %>% mutate(varname="date") ) %>%
                                   mutate(value_normalized= coalesce(value_normalized2) %>% as.character() ) %>% #we're
                                   mutate(varname_normalized= coalesce(varname_normalized2,varname_normalized) %>% as.character()  ) %>%
                                   dplyr::select(-value_normalized2,-varname_normalized2) %>% 
                                   filter(!(varname %in% "date" & is.na(value_normalized) ) ) %>% #as long as we keep this up first it'll be fine
                                   #Add in the dictionaries
                                   left_join(normalization_dictionaries %>%
                                               dplyr::select(varname,value,value_normalized2=value_normalized,varname_normalized2=varname_normalized ,value_qcode2=value_qcode ) 
                                             ) %>%
                                   mutate(value_normalized= coalesce(value_normalized2,value_normalized)) %>%
                                   mutate(value_qcode= coalesce(value_qcode2,value_qcode)) %>%
                                   mutate(varname_normalized= coalesce(varname_normalized2,varname_normalized)) %>%
                                   dplyr::select(-value_normalized2,-varname_normalized2, -value_qcode2) %>% 
                                   filter(!value_normalized %in% c("drop","not a location","nonspecific" )  ) %>%
                                   #Add in the actors
                                   left_join(dictionary_actors %>% rename(value_normalized2=value_normalized, value_qcode2=value_qcode)) %>%
                                   mutate(value_normalized= coalesce(value_normalized2,value_normalized)) %>%
                                   mutate(value_qcode= coalesce(value_qcode2,value_qcode)) %>%
                                   dplyr::select(-value_normalized2,-value_qcode2) %>% 
                                   filter(!value_normalized %in% "drop") %>%
                                   #Add in the reorders
                                   left_join(reorder_normalized %>% dplyr::select(crisno, email_id, input_crisis,varname,value,
                                                                                  sentence_number_int_aligned, value_normalized2=value_normalized) %>%
                                               mutate(value_normalized2= value_normalized2 %>% as.character() )) %>%
                                   mutate(value_normalized= coalesce(value_normalized2,value_normalized)) %>%
                                   dplyr::select(-value_normalized2) %>% 
  
                                   dplyr::mutate(value_normalized = strsplit(as.character(value_normalized), ";")) %>% 
                                   tidyr::unnest(value_normalized) %>%
                                   dplyr::mutate(value_normalized=value_normalized %>% trimws()) %>% 
                                   dplyr::mutate(crisno=as.numeric(crisno)) %>% 
                                   dplyr::mutate(value=tolower(value)) %>%
  
                                   mutate(varname_normalized=coalesce(varname_normalized, varname)) %>%
                                   mutate(value_normalized=coalesce(value_normalized, value)) %>%
                                   mutate(varname_normalized=varname_normalized %>% str_replace("_other","")) %>%  #We collapse other into just the normal flow here
                                   filter(!value_qcode %in% "drop") %>%
                                   filter(!value_normalized %in% "drop")
dim(icb_long_spans_clean_normalized) #953,335 #788,194 #835,319 #813,817 #815,679

# Normalize Dates and Sequencing


icb_long_spans_clean_normalized %>% saveRDS(paste0(here::here(), "/replication_data/out/ICBe_V1.1_long_clean.Rds")) 


```





Create variable "preceding_event"
Create a lead / lag that points to the previous event by group

Create variable "previous_event_timing"
```{r, eval=F}
alignment_key <- icb_wide_aligned_raw %>% 
  dplyr::ungroup() %>% 
  dplyr::select(email_id, crisno, sentence, raw_sentence_number_int, sentence_number_int_aligned, crisisno_crisis_version) %>% 
  dplyr::mutate(crisno = as.numeric(crisno))
```


Join in google sheet of tree simplification and drop variables that we replace

```{r message=FALSE, eval=F}
target_file <- paste0(here::here(),"/replication_data/in/icbe_litreview_trees_sentences.xlsx")
full_tree    <- readxl::read_excel(target_file, sheet="Codebook in Tree Form")


full_tree <- full_tree %>% dplyr::filter(!is.na(leaf) & leaf!='')

#The leafs aren't unique here because you can get to the same branch multiple ways

#This ignores conditions FYI
#Behavior # 3 bits
#think_aggression #1 bit
#say_aggression #1 bit
#do_aggression #1 bit
#Do_Armed_Unarmed #1 bit
#do_Act_Interact #1 bit
#full_tree$leaf_Simplified %>% unique() %>% length() #66 bits

#Only 70 bits total
  
#That's only another 29 bits

full_tree$leaf_think_simplified %>% unique() %>% length() #6
full_tree$leaf_say_simplified %>% unique() %>% length() #5
full_tree$leaf_do_simplified %>% unique() %>% length() #18

# full_tree$think_type_l0 %>% unique() %>% length() #6
# full_tree$say_type_l0 %>% unique() %>% length() #5
# full_tree$do_type_l0 %>% unique() %>% length() #18

```


```{r, eval=F}
icb_long_clean2 <- icb_long_clean %>% 
  dplyr::distinct() %>%
  dplyr::mutate(variable_normalized= variable_normalized %>%
                  stringr::str_replace("interact_location_other_clean","interact_location_defined")) %>%
  dplyr::mutate(variable_normalized= variable_normalized %>%
                  stringr::str_replace("condition_interact_location_other","condition_interact_location_defined")) %>% 
  dplyr::mutate(variable_normalized= variable_normalized %>%
                  stringr::str_replace("_other_clean","")) %>% 
  #We want to collapse others in with regulars
  dplyr::mutate(variable_normalized= variable_normalized %>% 
                  stringr::str_replace("_clean","")) %>% 
  #These only exist with actors now
  dplyr::mutate(variable_normalized= variable_normalized %>% 
                  stringr::str_replace("_other","")) %>% 
  #Only really matters for location_other
  dplyr::filter(value_normalized!='drop') %>%
  dplyr::filter(email_id!='expert3') %>% #expert3 Looks broken all the time
  #I'm going to move to a structure where event numbers can have a subsection that is condition rather than keeping it in the variable names
  dplyr::mutate(event_number_int_condition=variable_normalized %>%
                  stringr::str_detect('condition_') %>% as.numeric() ) %>%
  dplyr::mutate(variable_normalized=variable_normalized %>%
                  stringr::str_replace('condition_',''))  


# give every entry its own event_type var to control merging (avoid merging in details for actions that were 'thought about' or 'said')
# confirm that no email_crisis_sent_event have more than one event_type entry
icb_long_clean2 %>% dplyr::filter(variable_normalized == "event_type") %>% 
  dplyr::select("email_crisis_sent_event", "value_normalized") %>% 
  dplyr::rename(event_type=value_normalized) %>% dplyr::count(email_crisis_sent_event) %>% dplyr::filter(n>1)

icb_long_clean2 <- icb_long_clean2 %>% 
  dplyr::left_join(icb_long_clean2 %>% dplyr::filter(variable_normalized == "event_type") %>% 
                     dplyr::select("email_crisis_sent_event", "value_normalized") %>% 
                     dplyr::rename(event_type=value_normalized)) 



# icb_long_clean %>% dplyr::select(variable_normalized) %>% table()
# 
# icb_long_clean %>% dplyr::filter(variable_normalized=="interact_location") %>% dplyr::select(value_normalized) %>% table()
# icb_long_clean %>% dplyr::filter(variable_normalized=="condition_interact_location_other") %>% dplyr::select(value_normalized) %>% table()
# 
# icb_long_clean2 %>% dplyr::filter(variable_normalized=="interact_location") %>% dplyr::select(value_normalized) %>% table()
# icb_long_clean2 %>% dplyr::filter(variable_normalized=="interact_location_defined") %>% dplyr::select(value_normalized) %>% table()
# 
# temp <- icb_wide_aligned_raw %>% dplyr::select("do_actor_a", "do_actor_b", "condition_interact_location", "condition_interact_location_other") %>% dplyr::filter(!is.na(condition_interact_location_other))


```

```{r, eval=F}
temp_wide <- icb_long_clean2 %>%
             dplyr::select(icb_survey_version, email_id, input_crisis,  crisno, sentence_number_int_aligned, section, sentence,event_number_int,  event_number_int_condition, event_type, variable_normalized, value_normalized) %>%
             dplyr::mutate(value_normalized= value_normalized %>% trimws() %>% tolower()) %>%
             dplyr::mutate(variable_normalized= variable_normalized %>% trimws() %>% tolower()) %>%
             dplyr::filter(variable_normalized %in% (c( 'sentence_events', drops ) %>% trimws() %>% tolower()) ) %>%
             dplyr::distinct() %>% 
             dplyr::left_join(full_tree %>% 
                       dplyr::rename(value_normalized=leaf, variable_normalized=leaf_var)  %>% 
                       dplyr::mutate(value_normalized= value_normalized %>% trimws() %>% tolower()) %>% 
                       dplyr::mutate(variable_normalized= variable_normalized %>% trimws() %>% tolower())
                       ) %>% 
  dplyr::select(-event_type) #no longer needed after merge

temp_long <- temp_wide %>% 
         dplyr::select(-variable_normalized,-value_normalized) %>%
         tidyr::pivot_longer(-c(icb_survey_version, email_id,  input_crisis, crisno, sentence_number_int_aligned, section, sentence,   event_number_int,event_number_int_condition),
                      names_to='variable_normalized' , values_to= 'value_normalized')  %>%
         dplyr::filter(!is.na(value_normalized))

icb_long_clean_final <- icb_long_clean2 %>%
                   dplyr::select(icb_survey_version, email_id, input_crisis, crisno, sentence_number_int_aligned, section,  sentence,  event_number_int, event_number_int_condition, variable_normalized, value_normalized) %>%
                  rbind(temp_long) %>% dplyr::distinct() %>% 
                  dplyr::filter(!variable_normalized %in% drops)
                  
```
