==> replication_data/rmd/03_aggregation.Rmd <==
---
title: "03_aggregation"
output:
  html_document:
    df_print: paged
  toc: yes
editor_options:
  chunk_output_type: inline
---


# Intro

## Library Loads


```{r}
library(flextable)
library(tidyverse)

`%>%` <- magrittr::`%>%`

rex_paste <- function(x) { x %>% unique() %>% na.omit() %>% sort() %>% paste0(sep=";", collapse=";")}
rex_count <- function(x) { x %>% unique() %>% na.omit() %>% sort() %>% paste0(sep=";", collapse=";")}

```

## Hand coded aggregations

```{r}

icb_long_clean <- readRDS(file=paste0(here::here(), "/replication_data/out/ICBe_V1.1_long_clean.Rds")) %>%
                  filter(!varname %in% c("lastviewed","crisis",'sentencenumber','crisno', 'icb_survey_version',  'section' )) %>%
                  select(-sentence, -sentence_clean, -varname, -value, -status) %>% 
                  distinct()
dim(icb_long_clean)  
#icb_long_clean %>% count(varname_normalized) %>% arrange(desc(n)) %>% View()

coders_per_crisis <- icb_long_clean %>% dplyr::select(crisno,  email_id, expert) %>% distinct() #%>% dplyr::count(crisno)
coders_per_crisis_tally <- coders_per_crisis %>% mutate(novice=expert==0) %>% group_by(crisno) %>% summarise( expert=sum(expert==1), novice=sum(novice==1) )
crisis_sentences_per_coder <- icb_long_clean %>% filter(varname_normalized %>% str_detect("sentence_events")) %>% dplyr::count(email_id, crisno)


#icb_long_clean_original_393_43 <- icb_long_clean_original %>% filter(crisno==393 & sentence_number_int_aligned==43 ) %>% janitor::remove_empty() 
#Choose the most common sentence for each
sentences <- icb_long_clean %>% dplyr::count(crisno,sentence_number_int_aligned,sentence_span_text) %>% arrange(crisno, sentence_number_int_aligned, desc(n)) %>% 
              group_by(crisno, sentence_number_int_aligned) %>% filter(row_number()==1) %>% dplyr::select(-n) %>%
              ungroup()

crises <- icb_long_clean %>% arrange(crisno,sentence_number_int_aligned) %>% dplyr::select(input_crisis, crisno) %>% group_by(crisno) %>% filter(row_number()==1) %>% ungroup()

```

# Sentence Level Agreement

This is a calculation of token level agreement. It is the first part of Algorithm 1. 

```{r}

coders_per_crisis <- icb_long_clean %>% 
                              dplyr::select(crisno, sentence_number_int_aligned, email_id) %>%
                              distinct() %>%
                              count(crisno,email_id) %>% #just confirming that each coder put something for at least five sentences ( they all did)
                              mutate(expert=email_id %>% str_detect('expert') %>% as.integer() )  %>%
                              count(crisno, expert) %>%
                              pivot_wider(id_cols=crisno, values_from=n, names_from=expert) %>%
                              janitor::clean_names() %>% 
                              mutate(x0= x0 %>% replace_na(0) ) %>%
                              mutate(x1= x1 %>% replace_na(0) ) %>%
                              rename(coders_novice=x0, coders_expert=x1)

votes_per_crisis_sentence <- icb_long_clean %>%
                            mutate(expert=email_id %>% str_detect('expert') %>% as.integer() ) %>% #redundant now
                            #there are dupes here and I don't know why
                            dplyr::select(crisno,sentence_number_int_aligned, sentence_span_text, varname_normalized, value_normalized, expert, email_id) %>% distinct() %>%
                            count(crisno,sentence_number_int_aligned, sentence_span_text, varname_normalized, value_normalized, expert) %>%
                            pivot_wider(id_cols=c(crisno,sentence_number_int_aligned, sentence_span_text, varname_normalized, value_normalized), values_from=n, names_from=expert) %>%
                            janitor::clean_names() %>% 
                            mutate(x0= x0 %>% replace_na(0) ) %>%
                            mutate(x1= x1 %>% replace_na(0) ) %>%
                            rename(votes_novice=x0, votes_expert=x1) %>%
                            left_join(coders_per_crisis) %>%
                            mutate(percent_expert= (votes_expert/coders_expert) %>% replace_na(0)) %>%
                            mutate(percent_novice= (votes_novice/coders_novice) %>% replace_na(0)) %>%
                            mutate(votes_total=votes_expert+votes_novice) %>%
                            mutate(algo1_part1 = as.numeric(
                                     percent_expert>=0.5  #majority of experts
                                   )) %>%
                            mutate(algo1_part2 = as.numeric(
                                     votes_expert>0 & #At least one expert
                                     percent_novice>=0.5 #and majority of novices
                                   )) %>%
                            #If still no majority pick the expert aproved option with the most total votes
                            group_by(crisno,sentence_number_int_aligned, sentence_span_text, varname_normalized) %>%
                                mutate(algo1_part3 = as.numeric(  max(algo1_part1)==0 & 
                                                                    max(algo1_part2)==0 &
                                                                    votes_expert>0 & #at least one expert
                                                                    votes_total==max(votes_total) & #
                                                                    votes_total>1 #require at least 2 votes this is important
                                                                  ) ) %>% #no options were picked yet and this one has the most votes (at least 1 expert)
                            ungroup() %>%
                            mutate(algo1 = as.numeric( algo1_part1 | algo1_part2 | algo1_part3) )  %>%
                            arrange(crisno,sentence_number_int_aligned, sentence_span_text, varname_normalized, desc(algo1))

table(votes_per_crisis_sentence$algo1) #215619 #281663 #148732 #137153 #149,102


```

# Accepted Events

This is the second part of Algorithm 1. 

```{r}

icb_long_clean_accepted <- icb_long_clean %>% 
                           left_join(votes_per_crisis_sentence %>% dplyr::select(crisno,sentence_number_int_aligned,varname_normalized,value_normalized,algo1) ) %>%
                           filter(algo1==1) %>% 
                           filter(email_id %>% str_detect('expert')) #we already require at least one expert vote
dim(icb_long_clean_accepted) #200700 accepted tokens
#Now we need to collect the surviving tokens by the events they 
paste_rex=function(x) { x %>% str_split(";") %>% unlist()  %>% unique() %>% sort() %>% paste(collapse=";")}
icb_events <- icb_long_clean_accepted %>%
              dplyr::select(email_id, crisno, sentence_number_int_aligned, event_number_int, varname_normalized, value_normalized) %>% distinct() %>%
              arrange(email_id, crisno, sentence_number_int_aligned, event_number_int, varname_normalized, value_normalized) %>% 
              pivot_wider(id_cols=c( email_id, crisno, sentence_number_int_aligned, event_number_int), names_from=varname_normalized, values_from=value_normalized, values_fn=paste_rex) %>% #within a single event from that coder
              filter(!is.na(do_actor_a)    | 
                     !is.na(do_actor_b)    | 
                     !is.na(think_actor_a) | 
                     !is.na(say_actor_a)   | 
                     !is.na(say_actor_b) #remember to go back and collapse the others again
              ) %>%
              mutate(do_actor_b  = ifelse( do_actor_a==do_actor_b, '', do_actor_b ) ) %>% #There was a degenerate option to put the same actor in both A and B if it was acting alone
              mutate(say_actor_b = ifelse( say_actor_a==say_actor_b, '', say_actor_b ) ) %>%
              pivot_longer(cols=-c(email_id, crisno, sentence_number_int_aligned, event_number_int, think_actor_a,say_actor_a,say_actor_b, do_actor_a,do_actor_b), 
                           names_to="varname_normalized",
                           values_to="value_normalized",
                           values_transform=as.character) %>%
              dplyr::select(-email_id,-event_number_int) %>% #remove the person and event info
              pivot_wider(id_cols=c(  crisno, sentence_number_int_aligned, do_actor_a, do_actor_b, think_actor_a, say_actor_a ,say_actor_b),
                          names_from=varname_normalized, values_from=value_normalized, values_fn=paste_rex) %>% #collapse back down again
              full_join(sentences) %>%
              left_join(crises %>% rename(crisis_text=input_crisis) ) %>%
              relocate(sentence_span_text, .after = sentence_number_int_aligned) %>%
              relocate(crisis_text, .after = crisno) %>% 
              arrange( crisno, sentence_number_int_aligned) 
  
  
dim(icb_events) #15465 #17625 #18372 #17678 #14264 

```

## Events long

```{r}

#Qcodes
Qcodes <- icb_long_clean %>% dplyr::select(crisno, varname_normalized, value_normalized, value_qcode) %>% distinct() %>% filter(!is.na(value_qcode))

#This loses the structure of the diad, we need a new notion of an actor set
icb_events_long <- icb_events %>%
                   mutate(actor_set = paste0( think_actor_a,";",say_actor_a,";",say_actor_b,";",do_actor_a,";",do_actor_b)) %>%
                   pivot_longer(cols=-c(crisno, crisis_text, sentence_number_int_aligned, sentence_span_text, actor_set), 
                                 names_to="varname_normalized",
                                 values_to="value_normalized",
                                 values_transform=as.character) %>%
                   filter(!is.na(value_normalized)) %>% 
                   mutate(value_normalized = strsplit(as.character(value_normalized), ";")) %>% 
                   unnest(value_normalized) %>% 
                   distinct() %>% 
                   left_join(Qcodes) 

```


```{r, eval=T}

saveRDS(votes_per_crisis_sentence, paste0(here::here(), "/replication_data/out/ICBe_V1.1_long_agreement.Rds")) 
#readr::write_tsv(votes_per_crisis_sentence, paste0(here::here(), "/replication_data/out/ICBe_V1.1_long_agreement.tsv")) 

saveRDS(icb_long_clean, paste0(here::here(), "/replication_data/out/ICBe_V1.1_long.Rds")) 
#readr::write_tsv(icb_long_clean, paste0(here::here(), "/replication_data/out/ICBe_V1.1_long.tsv")) 

saveRDS(icb_events, paste0(here::here(), "/replication_data/out/ICBe_V1.1_events_agreed.Rds")) 
#readr::write_tsv(icb_events, paste0(here::here(), "/replication_data/out/ICBe_V1.1_events_agreed.tsv")) 

saveRDS(icb_events_long, paste0(here::here(), "/replication_data/out/ICBe_V1.1_events_agreed_long.Rds")) 
#readr::write_tsv(icb_events_long, paste0(here::here(), "/replication_data/out/ICBe_V1.1_events_agreed_long.tsv")) 


```


# Refactored Below


```{r, eval=F}

my.max <- function(x) ifelse( !all(is.na(x)), max(x, na.rm=T), NA)

keep_crisis=393
keep_sentence=38

options(warn = -1) #Groups that have no values generate max errors
#This doesn't have this yet
#event_number_int,  event_number_int_condition
rex_paste <- function(x){ paste0(x  %>% trimws() %>% sort() %>% unique() %>% na.omit(), collapse=";")  }
temp_long_list <- list()
for(crisis in unique(icb_long_clean$crisno)){
  print(crisis)
  
  temp_long_partial <- 
    icb_long_clean %>% 
      dplyr::filter(crisno == crisis) %>%
      #filter(crisno==keep_crisis & sentence_number_int_aligned==keep_sentence )  %>% #for debugging
      dplyr::select(email_id,crisno,sentence_number_int_aligned, section, sentence, event_number_int, event_number_int_condition,variable_normalized, value_normalized) %>% 
      dplyr::filter(value_normalized!='' & !is.na(value_normalized)) %>%
      dplyr::group_by(crisno,sentence_number_int_aligned, event_number_int_condition, variable_normalized, value_normalized) %>%
      dplyr::select(-event_number_int) %>%
      dplyr::distinct() %>%
      dplyr::mutate(
        email_id = email_id %>% as.factor(), 
        #crisno= crisno %>% as.factor(),
        #sentence_number_int_aligned= sentence_number_int_aligned %>% as.factor(),
        crisno_sentence_number_int_aligned= paste0(crisno, "_", sentence_number_int_aligned) %>% as.factor(), #make sure these are the only factors or it'll try to criss cross sentence number and crisno etc.
        section= section %>% as.character(),
        sentence= sentence %>% as.character(),
        variable_normalized= variable_normalized %>% as.factor()
      )  %>%
      dplyr::select(email_id , crisno,sentence_number_int_aligned, event_number_int_condition, variable_normalized, value_normalized) %>%
      distinct() %>%
      dplyr::group_by(email_id , crisno, sentence_number_int_aligned, event_number_int_condition, 
               variable_normalized,.drop=FALSE) %>% 
        #This drop false means every coder gets a '' whether they filled in that variable or not #section, sentence going to have to ignore the sentence text for now
        #You'll get a warning about adding it as a group
        summarise(value_normalized=value_normalized %>% rex_paste() ) %>%
      #that will add in dead ones now we want to split them back out
      dplyr::mutate(value_normalized = value_normalized %>% stringr::str_replace_all(";;",";")) %>% 
      dplyr::mutate(value_normalized = strsplit(as.character(value_normalized), ";", fixed=T)) %>% 
      tidyr::unnest(value_normalized) %>%
      dplyr::distinct() %>%
      #add a confidence score for every token
      dplyr::mutate(confidence=ifelse(variable_normalized=='raterconfidence', value_normalized, NA)) %>%
      dplyr::mutate(confidence=confidence %>% factor(levels=c('none','low','high','complete')) %>% as.numeric() ) %>%
      dplyr::mutate(confidence_expert= email_id %>% stringr::str_detect('expert') %>% ifelse(confidence,NA) %>% my.max() ) %>%
      dplyr::mutate(confidence_undergrad= email_id %>% stringr::str_detect('undergrad') %>% ifelse(confidence,NA) %>% my.max() )  %>%
  
      dplyr::group_by(crisno,sentence_number_int_aligned) %>%
        dplyr::mutate(
          total_coders=email_id %>% unique() %>% length(),
          total_coders_expert=email_id %>% unique() %>% stringr::str_detect('expert') %>% sum() ,
          total_coders_undergrad=email_id %>% unique() %>% stringr::str_detect('undergrad') %>% sum()
        ) %>%
      dplyr::ungroup() %>%
    
      #Selected By
      dplyr::group_by(crisno,sentence_number_int_aligned, event_number_int_condition, variable_normalized, value_normalized) %>%
        dplyr::mutate(
          selected_by_experts=email_id %>% unique() %>% stringr::str_detect('expert') %>% sum() ,
          selected_by_undergrads=email_id %>% unique() %>% stringr::str_detect('undergrad') %>% sum()
        ) %>%
      dplyr::ungroup() %>%
      dplyr::mutate(selected_by_any=selected_by_experts+selected_by_undergrads) %>%
      dplyr::left_join(sentences) %>%
      dplyr::mutate(
          selected_by_any_perc=selected_by_any/total_coders,
          selected_by_experts_perc=selected_by_experts/total_coders_expert ,
          selected_by_undergrads_perc=selected_by_undergrads/total_coders_undergrad ,
      ) %>%
      dplyr::ungroup() 

  temp_long_list[[as.character(crisis)]] <- temp_long_partial
}

temp_long <- dplyr::bind_rows(temp_long_list)
#table(temp_long$crisno) %>% length() 475
# for debugging
# try(if("icb_survey_version" %in% names(temp_long$icb_survey_version)) stop("icb_survey_version appeared..."))
    

```


```{r, eval=F}

#We have to compile confidence separately
sentence_confidence_scores <- temp_long %>% 
  dplyr::select(email_id,crisno,sentence_number_int_aligned, confidence,confidence_expert,confidence_undergrad) %>% distinct() %>% 
  mutate(confidence= confidence %>% replace(., !is.finite(.), NA),
         confidence_expert = confidence_expert %>% replace(., !is.finite(.), NA),
         confidence_undergrad = confidence_undergrad %>% replace(., !is.finite(.), NA)
         ) %>% 
  dplyr::select(-email_id) %>%
  group_by(crisno, sentence_number_int_aligned) %>%
  summarise_all(mean, na.rm=T) %>%
  mutate(confidence= confidence %>% replace(., !is.finite(.), NA),
         confidence_expert = confidence_expert %>% replace(., !is.finite(.), NA),
         confidence_undergrad = confidence_undergrad %>% replace(., !is.finite(.), NA)
         ) 

codings_long <- temp_long %>%
                mutate(
                       expert_majority=(selected_by_experts>(total_coders_expert/2) ) %>% as.integer() , #& selected_by_experts>=2 there are 31 crises with just 1 expert coder
                       undergrad_majority= (selected_by_undergrads > (total_coders_undergrad/2) & total_coders_undergrad>=2 ) %>% as.integer()  ,
                       expert_or_undergrad_majority= ( expert_majority | (undergrad_majority & selected_by_experts>=1) ) %>% as.integer() 
                ) %>%
                group_by(crisno,sentence_number_int_aligned, variable_normalized, value_normalized) %>%
                  mutate(
                        total_coders_expert=total_coders_expert %>% replace(., is.nan(.), NA) %>% mean(na.rm=T) %>% replace(., is.nan(.), NA),
                        total_coders_undergrad=total_coders_undergrad %>% replace(., is.nan(.), NA) %>% mean(na.rm=T) %>% replace(., is.nan(.), NA)
                  ) %>%
                ungroup() %>%
                dplyr::select(-email_id) %>%
                dplyr::select(-confidence,-confidence_expert,-confidence_undergrad) %>%
                distinct() %>% #bc expert majority is maxed now these rows should be fine
                left_join(sentence_confidence_scores) %>%
                arrange(crisno,sentence_number_int_aligned, event_number_int_condition, 
                        variable_normalized, expert_majority %>% desc(), undergrad_majority %>% desc(), undergrad_majority %>% desc(), selected_by_any %>% desc() ) #event_number_int


options(warn = 1)

```


```{r, eval=F}

#this has to be temp because we actually do a lot of filtering at the wide stage and so need to go back to long
codings_long_agreement <- codings_long %>%
                          #If picked by a majority of experts then in automatically
                          mutate(keep_majority_expert= 
                                    (expert_majority==1)   %>% as.numeric() #selected_by_experts>=2 & #selected by a majority of experts
                           ) %>%
                           #If picked by any expert and a majority of undergrads then keep
                           mutate(keep_majority_undergrad= (selected_by_experts>=1 & expert_majority==0 & undergrad_majority==1)  %>% as.numeric() ) %>%
                           #if either majority keep
                           mutate(keep_majority= (keep_majority_expert | keep_majority_undergrad) %>% as.integer() ) %>%
  
                           group_by(crisno, sentence_number_int_aligned, event_number_int_condition, variable_normalized) %>%
                                mutate(keep_majority_max=keep_majority %>% max(na.rm=T)) %>%
                                mutate(selected_by_any_max= (selected_by_any* (value_normalized!='')) %>% max(na.rmt=T) ) %>% #only consider non missing
                           ungroup() %>%
                           #If a question still doesn't achieve a majority then pick the answer that just recieved the most votes (and at least 1 expert vote)
                           mutate(keep_leftstanding = (keep_majority_max==0 & #Wasn't picked by a majority
                                                         selected_by_experts>=1 & #was picked by a least one expert
                                                         (value_normalized!='' & selected_by_any_max %in% selected_by_any) #Has the most total votes
                                                       ) %>% as.numeric() ) %>%
                           mutate(keep = ( keep_majority==1 | keep_leftstanding==1 ) %>% as.numeric() ) 

codings_long_agreement_393_43 <- codings_long_agreement %>% filter(crisno==393 & sentence_number_int_aligned==43 ) %>% janitor::remove_empty() 
codings_long_agreement_393_59 <- codings_long_agreement %>% filter(crisno==393 & sentence_number_int_aligned==59 )
codings_long_agreement_393_38 <- codings_long_agreement %>% filter(crisno==393 & sentence_number_int_aligned==38 )
codings_long_agreement_284_16 <- codings_long_agreement %>% filter(crisno==284 & sentence_number_int_aligned==16 )

codings_long_agreed <- codings_long_agreement %>%
                           filter(value_normalized!='')  %>%
                           filter(keep==1) %>% 
                           group_by(crisno, sentence_number_int_aligned) %>%
                              mutate(no_events= max( variable_normalized=="Event_Any" & value_normalized=="no") ) %>% #make sure you do the filter or this won't twork
                           ungroup()

dim(codings_long_agreed) #249,840     19  tokens that have agreement
codings_long_agreed %>% dplyr::select(keep_majority_expert, keep_majority_undergrad, keep_majority, keep_leftstanding) %>% summarise_all(sum)




(table(expert_majority=codings_long$expert_majority,
      undergrad_majority=codings_long$undergrad_majority) / nrow(codings_long) ) %>% round(2)
#               undergrad_majority
#expert_majority    0    1
#              0 0.40 0.04
#              1 0.40 0.16


```


```{r, eval=F}

#
#
#
#Maybe we use this as a screen on the original codings and then collect what survived to make subevents?
rex_paste <- function(x){ paste0(x %>% unlist() %>% trimws() %>% sort() %>% unique() %>% na.omit(), collapse=";")  }
codings_wide_agreed <- icb_long_clean  %>% arrange(crisno,sentence_number_int_aligned) %>%
                        #filter(crisno==59) %>%
                        dplyr::inner_join(codings_long_agreed) %>% #only keep ones that show up in both
                        filter(email_id %>% str_detect('expert') ) %>%
                        dplyr::select(email_id,crisno,sentence_number_int_aligned, event_number_int, event_number_int_condition, sentence,variable_normalized, value_normalized) %>%
                        distinct() %>%
                        pivot_wider(id_cols=c(email_id,crisno,sentence_number_int_aligned, sentence,event_number_int, event_number_int_condition), 
                                    names_from=variable_normalized, values_from=value_normalized, values_fn=rex_paste) %>%
                        distinct() %>%
                        group_by(crisno,sentence_number_int_aligned, event_number_int_condition, sentence, say_actor_a, say_actor_b, do_actor_a, do_actor_b) %>%
                          dplyr::summarise_all(rex_paste)
                        #filter(crisno==59 & sentence_number_int_aligned==3) 
codings_wide_agreed[codings_wide_agreed == ""]<-NA
#do the filters here after guaranteeing NAs

#Kill degenerate events
#We have a problem in that we want to reject certain kinds of events but we need to switch them to no events not drop the sentence all together
#This pruning basically works as intended but we need to go back and fill in null codings with a no event coding
codings_wide_agreed_pruned <- codings_wide_agreed %>% 
                              arrange(crisno,sentence_number_int_aligned) %>%
                              #filter(crisno==2 ) %>% #& sentence %>% str_detect('assassinated')
                              
                              #A number of things stop making sense at this level
                              dplyr::select(-behavior) %>%
                              dplyr::select(-event_type) %>%
                              dplyr::select(-event_any) %>%
                              dplyr::select(-event_count) %>%
                              dplyr::select(-raterconfidence) %>%
                              dplyr::select(-event_number_int) %>%
                              dplyr::select(-email_id) %>%
  
                              #mutate(reject_eventbutnobehavior=!event_any %in% "no" & is.na(behavior)) %>% #it's pre NA conversion below so ''
                              #there's a degenerate kind of actor group where behavior wasn't agreed upon by everyone but there's still some info. This effectively requires each sentence to only have one kind of event, think, say, do but the details might differ slightly by different sets of actors involved.   
                              #I don't know why but the actos are infact NAs, I guess because of the group
                              mutate(reject_degendactorsdo  = !is.na(do_actor_b) & is.na(do_actor_a)     )  %>% 
                              mutate(reject_degendactorssay = !is.na(say_actor_b) & is.na(say_actor_a)   )  %>% 
                              mutate(reject_degendactorsall = is.na(do_actor_b) & is.na(do_actor_a) & is.na(say_actor_b) & is.na(say_actor_a) & is.na(think_actor_a)  ) %>%  #I guess we're just going to reject background codings all together

                              mutate(reject_degenleafdo    = !is.na(do_actor_a) & is.na(leaf_do_original)          ) %>% #kill off pairings that have a do but no do actor a
                              mutate(reject_degenleafsay   = !is.na(say_actor_a) & is.na(leaf_say_original)        ) %>%
                              mutate(reject_degenleafthink = !is.na(think_actor_a) & is.na(leaf_think_original)    ) %>%
                                
                              mutate(reject_degenleafdo2   = is.na(do_actor_a) & !is.na(leaf_do_original)  ) %>% #kill of pairings that have a leaf but no actor
                              mutate(reject_degenleafsay2  = is.na(say_actor_a) & !is.na(leaf_say_original) ) %>%
                              mutate(reject_degenleafthink2= is.na(think_actor_a) & !is.na(leaf_think_original) ) %>%
                                
                              #arrange(crisno,sentence_number_int_aligned) %>%
                              #mutate(crisno_sent= paste0(crisno,"_",sentence_number_int_aligned)) %>%
                              #group_by(crisno_sent) %>%
                              #  mutate(event_any_max= max( !event_any %in% 'no' ) ) %>%
                              #  mutate(reject_nullevent = event_any_max>0 & event_any %in% c("no")  ) %>% #If any group has events then drop any groups without events
                              #ungroup() %>%
                              rowwise() %>%
                                mutate(reject= 
                                               #reject_eventbutnobehavior +
                                               reject_degendactorsdo +
                                               reject_degendactorssay +
                                               reject_degendactorsall +
                                               reject_degenleafdo +
                                               reject_degenleafsay +
                                               reject_degenleafthink +
                                               reject_degenleafdo2 +
                                               reject_degenleafsay2 +
                                               reject_degenleafthink2 #+
                                               #reject_nullevent
                                         ) %>% 
                              ungroup()

#codings_wide_agreed_pruned  %>% t() %>% View() #%>% filter(sentence_number_int_aligned==12 )
#rejects <- codings_wide_agreed_pruned %>% filter(reject>0 ) 
#dim(rejects)   #19346    60
#rejects %>% t() %>% View()
#codings_wide_agreed_pruned %>% filter(crisno==2 & sentence_number_int_aligned==11) 
#codings_wide_agreed %>% filter(crisno==2 & sentence_number_int_aligned==11) 

#codings_wide_agreed_384_43 <- codings_wide_agreed_pruned %>% filter(crisno==384 & sentence_number_int_aligned==43)  
#codings_wide_agreed_284_24 <- codings_wide_agreed_pruned %>% filter(crisno==284 & sentence_number_int_aligned==24)  
#codings_wide_agreed_447_14 <- codings_wide_agreed_pruned %>% filter(crisno==447 & sentence_number_int_aligned==14)  

#codings_wide_agreed_2_11 <- codings_wide_agreed_pruned %>% filter(crisno==2 & sentence_number_int_aligned==11) 

#codings_wide_agreed %>% filter(crisno==265 & sentence_number_int_aligned==28)  

codings_wide_agreed_pruned_final <- codings_wide_agreed_pruned %>% 
                                    filter(reject==0) %>% dplyr::select(-starts_with("reject")) %>%
                                    right_join(codings_wide_agreed_pruned %>% dplyr::select(crisno,sentence_number_int_aligned,event_number_int_condition,sentence) %>% distinct() ) %>% 
  
                                    #Generate final event ordering
                                    rowwise() %>%
                                    mutate(final_event_ordering=ifelse(!is.na(do_timing_reorder),
                                                                       #If more than one get past filter we take the smaller
                                                                       (do_timing_reorder %>% str_split(pattern=';', simplify = T) %>% unlist() %>% as.numeric() %>% min(na.rm=T)) +0.5,
                                                                       
                                                                       as.numeric(sentence_number_int_aligned) ) ) %>%
                                    ungroup() %>%
                                    arrange(crisno, final_event_ordering, sentence_number_int_aligned) %>% relocate(final_event_ordering, .after = sentence_number_int_aligned) %>%
                                    group_by(crisno) %>%
                                    mutate(final_event_ordering=row_number()) %>%
                                    ungroup()
        
  
```


```{r, eval=F}

library(glue)

glue_rex <- function(x) glue(x, .open = "{|",.close = "|}", .envir = parent.frame()) #escape variables in the more complicated {||} syntax
target_file <- paste0(here::here(),"/replication_data/in/icb_manual_recoding_master_sheet.xlsx")
dictionary_actors_labels    <- readxl::read_excel(target_file, sheet="actors") %>% dplyr::select(crisno, value_normalized=value_normalized_wikidata_id, value_normalized_label) %>% dplyr::distinct() %>% na.omit()
convert_qcode_to_english <- function(crisis, qcodes){
    crisis_actors <- dictionary_actors_labels %>% filter(crisno==crisis)
    crisis_actors_set <- crisis_actors$value_normalized_label
    names(crisis_actors_set) <- crisis_actors$value_normalized
    actors_english <- crisis_actors_set[qcodes %>% str_split(";") %>% unlist()]  %>% na.omit() %>% paste0(collapse=";")
    if(length(actors_english)==0){actors_english=NA}
    return(actors_english)
}
codings_wide_agreed_pruned_labeled <- NULL
codings_wide_agreed_pruned_labeled <-  codings_wide_agreed_pruned_final %>% 
  mutate(crisno_sent=paste0(crisno,'.',sentence_number_int_aligned)) %>%
  rowwise() %>%
  mutate(say_actor_a_labels=  convert_qcode_to_english(crisno, say_actor_a)) %>%
  mutate(say_actor_b_labels=  convert_qcode_to_english(crisno, say_actor_b)) %>%
  mutate(do_actor_a_labels=  convert_qcode_to_english(crisno, do_actor_a)) %>%
  mutate(do_actor_b_labels=  convert_qcode_to_english(crisno, do_actor_b)) %>%
  mutate(think_actor_a_labels=  convert_qcode_to_english(crisno, think_actor_a)) %>%
  mutate(think_sentence=ifelse(
      !is.na(leaf_think_original) & !is.na(think_sentence_events) , 
      glue_rex("[{|think_actor_a_labels|}]{.underline text.color='black'} experienced a 
                          [{|leaf_think_original|}]{.underline text.color='black'} about 
                          [{|think_sentence_events|}]{.underline text.color='black'}."),
            glue_rex("[{|think_actor_a_labels|}]{.underline text.color='black'} experienced a 
                          [{|leaf_think_original|}]{.underline text.color='black'}.")
  )) %>%
  mutate(say_sentence=ifelse(
      !is.na(leaf_say_original) , 
      glue_rex("[{|say_actor_a_labels|}]{.underline text.color='black'} initiated
                        [{|leaf_say_original|}]{.underline text.color='black'} toward  
                        [{|say_actor_b_labels|}]{.underline text.color='black'}.") ,
      ''
  )) %>%

  mutate(consequence_sentence=ifelse(
    consequence %in% "will happen" , 
    glue_rex("That the following will [happen]{.underline text.color='black'}.") ,
    glue_rex("That if the following [won't happen]{.underline text.color='black'}.")
  )) %>%

  mutate(condition_sentence=ifelse(
    condition %in% "happens" , 
    glue_rex("If the following [happens]{.underline text.color='black'}.") ,
    glue_rex("If the following [does not happen]{.underline text.color='black'}.")
  ))  %>%
  
  mutate(do_sentence=ifelse(
    !is.na(leaf_do_original) & !is.na(do_actor_b), 
    glue_rex("[{|do_actor_a_labels|}]{.underline text.color='black'} 
              initiated 
              [{|leaf_do_original|}]{.underline text.color='black'}
              toward 
              [{|do_actor_b_labels|}]{.underline text.color='black'}." ) ,
    glue_rex("[{|do_actor_a_labels|}]{.underline text.color='black'} 
              initiated 
              [{|leaf_do_original|}]{.underline text.color='black'}")
   )) %>%

   mutate(codings_sentence=NA) %>%

   #
   mutate(codings_sentence=ifelse( is.na(leaf_say_original) & is.na(leaf_do_original) & is.na(leaf_think_original) & is.na(consequence) & is.na(condition) & event_number_int_condition %in% 0,
                                   glue_rex("No events.")
                                   ,codings_sentence)) %>%

   mutate(codings_sentence=ifelse( is.na(leaf_say_original) & is.na(leaf_do_original) & !is.na(leaf_think_original) & is.na(consequence) & is.na(condition) & event_number_int_condition %in% 0,
                                   glue_rex("{|think_sentence|}")
                                   ,codings_sentence)) %>%
     
    mutate(codings_sentence=ifelse( !is.na(leaf_say_original) & is.na(leaf_do_original) & is.na(leaf_think_original) & is.na(consequence) & is.na(condition) & event_number_int_condition %in% 0,
                                   glue_rex("{|say_sentence|}")
                                   ,codings_sentence)) %>%    
     
    mutate(codings_sentence=ifelse( is.na(leaf_say_original) & !is.na(leaf_do_original) & is.na(leaf_think_original) & is.na(consequence) & is.na(condition) & event_number_int_condition %in% 0,
                                   glue_rex("{|do_sentence|}")
                                   ,codings_sentence)) %>%     

    #Condition sentence
    mutate(codings_sentence=ifelse( is.na(leaf_say_original) & !is.na(leaf_do_original) & is.na(leaf_think_original) & is.na(consequence) & !is.na(condition) & event_number_int_condition %in% 0,
                                   glue_rex("[Condition] {|do_sentence|}")
                                   ,codings_sentence)) %>%     
  
    mutate(codings_sentence=ifelse( is.na(leaf_say_original) & !is.na(leaf_do_original) & !is.na(leaf_think_original) & is.na(consequence) & is.na(condition) & event_number_int_condition %in% 0,
                                   glue_rex("{|think_sentence|} {|do_sentence|}")
                                   ,codings_sentence)) %>%     
  
    mutate(codings_sentence=ifelse( !is.na(leaf_say_original) & is.na(leaf_do_original) & !is.na(leaf_think_original) & is.na(consequence) & is.na(condition) & event_number_int_condition %in% 0,
                                   glue_rex("{|think_sentence|} {|say_sentence|}")
                                   ,codings_sentence)) %>%        
     
    mutate(codings_sentence=ifelse( !is.na(leaf_say_original) & !is.na(leaf_do_original) & !is.na(leaf_think_original) & is.na(consequence) & is.na(condition)  & event_number_int_condition %in% 0,
                                   glue_rex("{|think_sentence|} {|say_sentence|} {|do_sentence|}")
                                   ,codings_sentence)) %>%        

    mutate(codings_sentence=ifelse( !is.na(leaf_say_original) & !is.na(leaf_do_original) & is.na(leaf_think_original) & is.na(consequence) & is.na(condition) & event_number_int_condition %in% 0,
                                   glue_rex("{|say_sentence|} {|do_sentence|}")
                                   ,codings_sentence)) %>%      
    
   mutate(codings_sentence=ifelse( !is.na(leaf_say_original) & !is.na(leaf_do_original) & is.na(leaf_think_original) & !is.na(consequence) & is.na(condition) & event_number_int_condition %in% 0,
                                   glue_rex("{|say_sentence|} {|consequence_sentence|} {|do_sentence|}")
                                   ,codings_sentence))  %>%

    mutate(codings_sentence=ifelse( !is.na(leaf_say_original) & !is.na(leaf_do_original) & is.na(leaf_think_original) & !is.na(consequence) & !is.na(condition) & event_number_int_condition %in% 0,
                                   glue_rex("{|say_sentence|} {|consequence_sentence|} {|do_sentence|} {|condition_sentence|}")
                                   ,codings_sentence))    
     



#codings_wide_agreed_labeled %>% dplyr::select(codings_sentence, contains('actor'), contains('_original'), consequence, condition, event_number_int_condition) #%>% View()

#codings_wide_agreed_labeled %>% dplyr::select(contains('actor'), contains('_original'), consequence, condition, event_number_int_condition, contains('_sentence')) #%>% View()

(codings_wide_agreed_pruned_labeled_2_11 <- codings_wide_agreed_pruned_labeled %>% filter(crisno==2, sentence_number_int_aligned==11) %>% 
                                    dplyr::select(crisno, sentence_number_int_aligned, contains('actor'), contains('_original'), consequence, condition, event_number_int_condition, contains('_sentence')) ) #%>% View()

(codings_wide_agreed_pruned_labeled_28_5 <- codings_wide_agreed_pruned_labeled %>% filter(crisno==28, sentence_number_int_aligned==5) %>% 
                                    dplyr::select(crisno, sentence_number_int_aligned, contains('actor'), contains('_original'), consequence, condition, event_number_int_condition, contains('_sentence')) ) #%>% View()

  
(codings_wide_agreed_pruned_labeled_34_17 <- codings_wide_agreed_pruned_labeled %>% filter(crisno==34, sentence_number_int_aligned==17) %>% 
                                    dplyr::select(crisno, sentence_number_int_aligned, contains('actor'), contains('_original'), consequence, condition, event_number_int_condition, contains('_sentence')) ) #%>% View()


```

# Validation Tests

```{r, eval=F}

keep_crisis=393
keep_sentence=38

(a <- icb_long_clean_original %>% filter(crisno==keep_crisis & sentence_number_int_aligned==keep_sentence ) %>% janitor::remove_empty()  ) #%>% View()
(b <- temp_long %>% filter(crisno==keep_crisis & sentence_number_int_aligned==keep_sentence ) %>% janitor::remove_empty()  ) #%>% View()
(c <- temp_wide %>% filter(crisno==keep_crisis & sentence_number_int_aligned==keep_sentence ) %>% janitor::remove_empty() )#%>% t()  ) #%>% View()
(d <- icb_long_clean %>% filter(crisno==keep_crisis & sentence_number_int_aligned==keep_sentence ) %>% janitor::remove_empty()  ) #%>% View()
(e <- codings_long %>% filter(crisno==keep_crisis & sentence_number_int_aligned==keep_sentence ) %>% janitor::remove_empty()  ) #%>% View()
(f <- codings_long_agreement %>% filter(crisno==keep_crisis & sentence_number_int_aligned==keep_sentence ) %>% janitor::remove_empty()   ) #%>% View()
(g <- codings_long_agreed %>% filter(crisno==keep_crisis & sentence_number_int_aligned==keep_sentence ) %>% janitor::remove_empty()   ) #%>% View()
(e <- codings_wide_agreed  %>% filter(crisno==keep_crisis & sentence_number_int_aligned==keep_sentence )  %>% janitor::remove_empty()  ) %>% t() #%>% View()

```
